Move the rest of CommunicationUtilities.cs into Framework
