+++ src/Build.UnitTests/BackEnd/TaskHostNodeKey_Tests.cs
@@ -1,6 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using Microsoft.Build.BackEnd;
 using Microsoft.Build.Internal;
 using Shouldly;
 using Xunit;
