+++ src/Framework/ErrorUtilities.cs
@@ -64,6 +64,17 @@ namespace Microsoft.Build.Framework
 
         /// <summary>
         /// Throws InternalErrorException.
+        /// Indicates the code path followed should not have been possible.
+        /// This is only for situations that would mean that there is a bug in MSBuild itself.
+        /// </summary>
+        [DoesNotReturn]
+        internal static void ThrowInternalErrorUnreachable()
+        {
+            throw new InternalErrorException("Unreachable?");
+        }
+
+        /// <summary>
+        /// Throws InternalErrorException.
         /// This is only for situations that would mean that there is a bug in MSBuild itself.
         /// </summary>
         [DoesNotReturn]
